Fix incorrect test for inverse transition function strict property
The test was meant to ensure that if the transfer function was strict that the
inverse transition function (if present) was also strict. This caused some regression
tests to fail, these tests turned out to be incorrect. In passing also changed
error messages for the strict tests in create aggregate to be the same as the
message which will be shown at run time for the same error.