Cleaned up xaggr.sgml.
I've tried to shorten things a bit - it certainly seems sensible
to discuss the case of inverse transition functions and floating-point
type, but that page is about aggregates, not floating-point
arithmetic.

I also didn't like that the previous text recommended creating your
own float aggregate "if you can live with the roundig anomalies".
That suggests that the paragraph about explains all the problems
that can arise, but it doesn't, and it shouldn't. People who
understand enough about floating-point arithmetic to make an
informed decision won't need such a suggestion, and we shouldn't
encourage those who don't to play it fast and loose.

This concludes my work on the documentation, or at least so I hope.