string_agg() needs to track the number of non-NULL strings.
string_agg() ignores rows where the 1st arg (the string) is NULL, but
treats a NULL 2nd arg (the delimiter) as an empty string. That makes it
impossible to mark the inverse transition function as strict, so we need
to track the number of non-NULL inputs to correctly return NULL if all
input strings where NULL.

This also fixed the regression test to correctly show '-' when
string_agg(bytea) return NULL - previoulsy it showed \x2d, which is
confusing.

Also updated window.out since this changes the result of the regression
test of string_agg() - what it returns now is hopefully right.