Make enabling JIT more granular
Previously, whether to JIT or not to JIT, was decided based on the total
cost of the plan.  If we reached the jit_above_cost for the plan then we
would do JIT compilation for all plan nodes.  Suddenly enabling JIT like
this when the plan reached the given cost could lead to large performance
penalties for large and complex plans.  A particularly bad case was
queries which access a partitioned table with a large number of
partitions. The time spent JITing such plans could often be lengthy.

Here we change things so that instead of enabling JIT on the entire plan,
we instead only enable JIT if a given plan node reaches the JIT cost
threshold.  This means that cheap leaf nodes, for example, partitions that
don't have many rows can be executed without JIT, but more costly parts of
the plan can still get JITed.

Since plan nodes may be executed more than once, for example, the inner
side of a Nested Loop join, here we add infrastructure into createplan.c
to calculate the number of times we expect to (re)invoke a given plan
node. The reason for this is that we want to take the number of times a
given node will be rescanned into account when considering JIT. We
multiply the total_cost of the plan node by the estimated calls and
compare that to the given JIT GUC in order to decide what/how we JIT
compile.

This changes the meaning of the jit_above_cost, jit_inline_above_cost and
jit_optimize_above_cost GUCs.  Users upgrading from previous versions may
want to reconsider their settings for these GUCs, especially so if they
increased them from their default values.

Whether to inline and optimize during JIT compilation is decided globally
per query rather than at the individual plan level.  This is because all
the code is sent to the JIT compiler to be compiled at once.  Inlining and
optimizing is enabled for all JIT enabled nodes if any plan node reaches
the required cost to enable the given option.