Allow sort order to be preserved through certain sql function calls
When calling functions such as date_trunc(..., column), the sort order of
the return values from the function will be the same as the "column"
argument.  The function only gets rid of some precision.

Unfortunately, if a given Path in the planner had PathKeys for "column",
then the planner would be unaware that those Pathkeys would suit anything
which required Pathkeys on date_trunc(..., column).  The planner would end
up adding an extra sort in cases such as:

SELECT date_trunc(..., column), sum(val) from tab group by 1 order by 1;

even if there was a btree index on tab (column).

Here we add a concept called super-pathkeys which allows a Pathkey to have
a parent PathKey which the order can be derived from.  In this case the
date_trunc(..., column) would be the Pathkey and "column" would be a
super Pathkey.  When matching up pathkeys, if an exact matching pathkey is
not found and a super pathkey exists for the pathkey, then the planner
will try to match that up.

This feature can give a good performance boost to timeseries type queries
which aggregate large amounts of data.  Often there is already an index on
a timestamp column.  Performing a group aggregate on a function such as
date_trunc would have required either hash aggregate or sorting the input
on the date_trunc function call.  When the number of rows was large then
sorting would never be that appealing.

This is unfinished and needs lots of comments improved.  It's also
currently not working as pathkeys_count_contained_in needs to know about
super pathkeys and I don't have the energy right now to code that.

Commit message written when very tired.