Allow simplehash users to invent ways of storing the hash status
Prior to this, simplehash.h required the user of simplehash to include a
field named "status" for the hash bucket struct, however that may have
been inconvenient for some as it could cause the struct to be larger than
required.  Here we allow custom macros to be defined to override the
default use of the "status" field.  Callers could do things like encode
the status in the alignment bits in a pointer field, for example.

 #define SH_STATUS_IS_EMPTY(e) ((((uintptr_t) (e)->data) & 1) == 0)
 #define SH_STATUS_IS_IN_USE(e) ((((uintptr_t) (e)->data) & 1) == 1)
 #define SH_STATUS_SET_EMPTY(e) (e)->data = (<type>) (((uintptr_t) (e)->data) & ~((uintptr_t) 1))
 #define SH_STATUS_SET_IN_USE(e) (e)->data = (<type>) (((uintptr_t) (e)->data) | ((uintptr_t) 1))

Callers using such methods would need to be careful to always unset the
status bit in the ->data field when they want to dereference the pointer.