Add Bitmapset indexes for faster lookup of EquivalenceMembers
For queries containing a large number of joins or for queries to
partitioned tables, the looking up of EquivalenceMembers in an
EquivalenceClass can become slow due to the large number of members.
Until now, the searching for EquivalenceMembers has been done using a
linear search over the EquivalenceClass's list of members. Here we
effectively add an index to allow faster lookups of these members.  This
is done by way of adding all members to all classes into a List in
PlannerInfo, then we have a Bitmapset in each EquivalenceClass with
members for each index in PlannerInfo's list which belong to this class.
Additionally, each RelOptInfo also maintains another Bitmapset which
stores each index in PlannerInfo's list of EquivalenceMember for all
members, in all classes, which mention this relation.  This allows us to
quickly find all EquivalenceMembers for a given RelOptInfo and a given
EquivalenceClass by intersecting these two lists.

When searching for members belonging to join rels, we must either
intersect or union all base RelOptInfo's member indexes then intersect the
result to the EquivalenceClass's indexes.  Whether we intersect or union
depends on if we need members that mention all relations or members that
mention any of the base relations making up the join relation.

This method of indexing can significantly speed up the planner when
querying a partitioned table with a large number of partitions when the
query contains a few joins.  Tests have shown around a x10 increase in
performance with 1000 partitions.