Have the planner try to push the ORDER BY Sort below WindowAggs
When the query contains a WindowFunc which the required sort order is some
prefix of the query's ORDER BY clause, instead of performing one Sort for
the WindowAgg and then later perform an Incremental Sort to further sort
the records into the final required order, we may as well push the full
sort down below the WindowAgg.  This means we have a single Sort node in
the plan rather than a Sort node and an additional Incremental Sort node
for the ORDER BY clause.

There are a few cases where we don't apply this optimization:

1. When the query has a DISTINCT clause.  This is evaluated after
WindowAggs and before the ORDER BY, so there's no point in doing any more
complex sorts that could be destroyed by Hash Aggregate.

2. If the query has a LIMIT clause then the top-level sort will be a top-n
sort.  In this case, it may be faster to perform the top-n sort than to
sort all the records with the additional sort columns of the ORDER BY
clause, so we just don't do it in this case.

3. If the top-level WindowAgg has a runCondition then that may filter out
rows.  If we were to push the ORDER BY's sort below the WindowAgg then
we'd have to sort rows that the runCondition may filter.  Since we have no
means to estimate how many rows will be filtered by the runCondition, it
seems safer not to try to optimize this case.

4. If the ORDER BY contains any WindowFuncs then we can't push the sort.
We must wait until the WindowAgg evaluates these before we can perform the
sort.

Author: Ankit Kumar Pandey <itsankitkp@gmail.com>
Author: David Rowley <dgrowleyml@gmail.com>
Reviewed-by: David Rowley <dgrowleyml@gmail.com>
Discussion: https://postgr.es/m/83d80853-a45c-d85c-68eb-59acfe7fb5fb@gmail.com