Adjust nodeUnique to use the same slot type as its subnode
When the Unique node's subnode uses TTSOpsVirtual the Unique node
previously had to form a MinimalTuple.  Forming the MinimalTuple was
costly and even more costly when that MinimalTuple needed to be deformed
again when comparing it to the next tuple.

Here we adjust the slot type of nodeUnique to make it use the same slot
type as its subnode.  In many cases this will be TTSOpsVirtual which can
be much cheaper to copy, especially so if there are only byval types in
the slot.  TTSOpsVirtual slots don't require any tuple deforming when
executing for the next tuple to compare it to the previous tuple.

Author: Denis Smirnov
Reviewed-by: Heikki Linnakangas, David Rowley
Discussion: https://postgr.es/m/388BA879-CAEC-4930-A4EE-F06DC1F12C96%40gmail.com