Optimize tuple comparisons in various executor nodes
In the Group, Limit (WITH TIES) and Unique executor nodes we have to
compare the current tuple with some previously saved tuple.  This is
done by evaluating an ExprState.  Before this commit that ExprState was
always initialized with tuple deforming steps.  The deform steps are
always required when the input slot is anything apart from
TTSOpsVirtual.  Here we adjust the code so that the deform steps are
skipped if any of the input slots are guaranteed to be TTSOpsVirtual.
This should speed up the tuple comparisons done in these node types when
the input slot is TTSOpsVirtual.

Author: David Rowley