Make unsetting the isReset flag the MemoryContext's job
Up until now, setting this boolean flag to false has been the job of the
allocation function in mcxt.c.  Making this the job of the MemoryContext
implementation can make things more efficient as adjusting the flag can
just be put in code paths that are only possible to reach in cases where
we may not have already done an allocation since the context was created
or reset.

Discussion: https://postgr.es/m/20240222225343.7a57fiv7quehhlur@awork3.anarazel.de