Improve performance of the generation memory context
This is a very simple change.  All we do is move the GenerationBlock's
context field so it's the first member of the struct.  This field is the
first one accessed in pfree() and repalloc(), making it first means no
bytes need to be added to the GenerationBlock's address in order to access
the stored in the context address.

THIS DOES NOT SEEM TRUE?!  blocks->nfree += 1; looks like the first
field accessed