Move TupleDesc.attrs out of line
Change TupleDesc attrs field so that it's no longer memory allocated at
the end of the TupleDesc struct.  Here we change this so that 'attrs' is
a pointer that always points to memory beyond the end of the struct.

This is just refactoring work to make way for a follow-on patch which
adds another variable length array to TupleDesc.