Optimize alignment calculations in tuple form/deform
This converts TupleDescDeformAttr.attalign from a char which is directly
derived from pg_attribute.attalign into a uint8 which specifies the
number of bytes to align the column by.  This removes the complexity of
checking each char value and transforming that into the appropriate
alignment call.  This can just be a simple TYPEALIGN passing in the
number of bytes.