Adjust EXPLAIN's output for disabled nodes
c01743aa4 added EXPLAIN output to display the plan node's disabled_node
count whenever that count is above 0.  Seemingly, there weren't many
people who liked that output as each parent of a disabled node would
also have a "Disabled Nodes" output due to the way disabled_nodes is
accumulated towards the root plan node.  It was often hard and sometimes
impossible to figure out which nodes were disabled from looking at
EXPLAIN.  You might think it would be possible to manually add up the
numbers from the "Disabled Nodes" output of a given node's children to
figure out if that node has a higher disabled_nodes count, but that
wouldn't have worked for Append and Merge Append nodes if some of the
children were run-time pruned during init plan. Those are not
displayed in EXPLAIN.

Here we attempt to improve this output by only showing "Disabled: true"
against only the nodes which are explicitly disabled themselves, which
seems to be the most desired output.  This is achieved by simply
checking if the given node has a higher disabled_nodes count than the
sum of its child nodes.

This also fixes a bug in make_sort() which was neglecting to set the
Sort's disabled_nodes field.  With the new output, the choice to not
maintain that field properly was clearly wrong as the disabled-ness of
the node was attributed to the parent instead.

Discussion: https://postgr.es/m/9e4ad616bebb103ec2084bf6f724cfc739e7fabb.camel@cybertec.at