Optimize nodeSort.c code
This adjusts nodeSort.c's code so that the code that only executed once
is moved out of line of the main ExecSort function.  This reduces the
amount of code in the common code path so that the code that's more
likely to be executed next comes next in the compiled binary.

Also, adjust ExecSort to split it into two functions, one to perform
Datum sorts and another version to support tuple sorting.  Writing the
code this way saves having to check which type of sort we're doing once
for each tuple we fetch from ExecSort.