Eliminate integer divide in fastpath locking code
FAST_PATH_REL_GROUP() was using integer division to determine the
fastpath locking group slot to use for the relation.  While this doesn't
seem to be very hot code, divide is a very slow operation to put in the
fastpath locking code.  In any case, the global variable we're dividing
by is always a power-of-two, so adjust the macro to use a bitwise-AND
rather than a divide.

In passing adjust the code that's setting FastPathLockGroupsPerBackend
so that it's more clear that the value being set is a power-of-two.
Also adjust some comments in the area which contained some magic
numbers.  This seemed better placed in the location where the #defines
were located rather than were they were being used.