Adjust ExecInit*() to initialize plans depth-first
This attempts to reduce memory fragmentation caused during executor
startup, where nodes will allocate some memory for themselves, then go
off and initialize their child nodes before allocating more memory for
themselves.  This causes the memory belonging to a single node to be
spread out more than what is needed.

Here we adjust the allocation order so children are generally initialized
first.  For nodes which can have an arbitrary number of child nodes,
such as Append and MergeAppend, the array to store the state of the
children must be allocated first, so in this case it's not possible to
initialize the children first as there is nowhere to store the child
states.  With many of these nodes, the child nodes are initialized last,
so there's no additional fragmentation, though the memory order isn't
the same as if we'd initalized them first.  It's likely this does not
matter much.