Adjust ExecInit to initialize plans depth-first
This attempts to reduce memory fragmentation caused during executor
startup where nodes will allocate some memory for themselves then go off
and initialize their child nodes before allocating more memory for
themselves.  This causes the memory belonging to a single node to be
spread out more than what is needed.

Here we adjust the allocation order so children are always initialized
first.  I didn't quite get every location as some nods such as Append
and MergeAppend need an array to store their child PlanStates and that
requires allocation.  In other places that require just an outer and
inner node, we can temporarily store those in local variables before we
allocate memory for the node itself.

This is an experiment. Not a complete patch